Added globals file, moved quiz result tallying there
