+++ components/quiz/QuizCard.js
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { Text, View, StyleSheet, TouchableOpacity, ScrollView } from 'react-native';
 import firebase from '../../Firebase';
+import h from '../../globals';
 
 import Response_Checkbox from './Response_Checkbox';
 import Response_Radio from './Response_Radio';
@@ -62,9 +63,9 @@ export default function QuizCard(props) {
       // this.setState({ modalVisible: true });
 
       // Tally the output flags, filter for the three highest, and randomize them
-      let outputFlags = tallyOutputFlags(quizEffects);
-      let topThree = getHighsAndLows(outputFlags, 3, 0)[0];
-      let suggestions = randomizeSuggestions(topThree);
+      let outputFlags = h.tallyOutputFlags(quizEffects);
+      let topThree = h.getHighsAndLows(outputFlags, 3, 0)[0];
+      let suggestions = h.randomizeSuggestions(topThree);
       // Set the outputSuggestions object with the randomized suggestions
       setOutputSuggestions({
         suggestion_1: hitpause.suggestions[suggestions[0]],
@@ -80,81 +81,6 @@ export default function QuizCard(props) {
       });
       setModalVisible(true);
     }
-  }
-
-  function tallyOutputFlags(flags) {
-    let outputFlags = {};
-    let modifiers = [];
-    // For each object in the array of flags
-    for (const key in flags) {
-      // For each property in the flag object
-      for (const flagKey in flags[key]) {
-        // If we are affecting the highest flags
-        if (flagKey.includes('_highest_')) {
-          let count = parseInt(flagKey.replace('_highest_', ''));
-          modifiers.push({ type: 'high', count: count, amount: parseFloat(flags[key][flagKey]) });
-        }
-        // If we are affecting the lowest flags
-        else if (flagKey.includes('_lowest_')) {
-          let count = parseInt(flagKey.replace('_lowest_', ''));
-          modifiers.push({ type: 'low', count: count, amount: parseFloat(flags[key][flagKey]) });
-        }
-        // If flags of this type already exist, sum them
-        else if (Object.keys(flags).includes(flagKey)) {
-          outputFlags[flagKey] = parseFloat(flags[flagKey]) + parseFloat(flags[key][flagKey]);
-        }
-        // Otherwise, add normally
-        else outputFlags[flagKey] = parseFloat(flags[key][flagKey]);
-      }
-    }
-
-    // For all of our special cases
-    for (const key in modifiers) {
-      let modifiedFlags = {};
-      // Depending on type of case, grab relevant flags
-      if (modifiers[key].type == 'high') {
-        modifiedFlags = getHighsAndLows(outputFlags, modifiers[key].count, 0)[0];
-      } else if (modifiers[key].type == 'low') {
-        modifiedFlags = getHighsAndLows(outputFlags, 0, modifiers[key].count)[1];
-      }
-      // Apply changes to those flags
-      for (const flagKey in modifiedFlags) modifiedFlags[flagKey] = modifiedFlags[flagKey] + modifiers[key].amount;
-      outputFlags = { ...outputFlags, ...modifiedFlags };
-    }
-    return outputFlags;
-  }
-
-  // TODO: Incomplete... will probably want to move this into summary screen
-  // TODO: Ties are unfair, as lower keys will always be chosen... need a way to randomly select when there are ties
-  function getHighsAndLows(flags, numHighs, numLows) {
-    let highValues = [], lowValues = [];
-    let sortedFlags = Object.entries(flags).sort((a, b) => (a[1] < b[1]));
-    if (numHighs > 0) {
-      highValues = sortedFlags.slice(0, numHighs);
-    }
-    if (numLows > 0) {
-      lowValues = sortedFlags.slice(sortedFlags - 1 - numLows, sortedFlags.length - 1);
-    }
-    return [Object.fromEntries(highValues), Object.fromEntries(lowValues)];
-  }
-
-  function randomizeSuggestions(flags) {
-    let n = 0;
-    for (const key in flags) {
-      let squaredDoubleFlag = (flags[key] * 2) ** 2;
-      flags[key] = squaredDoubleFlag + n;
-      n = squaredDoubleFlag + n;
-    }
-    let randomInt = Math.floor(Math.random() * n);
-    for (const key in flags) {
-      if (flags[key] > randomInt) {
-        delete flags[key];
-        // TODO: Verify this actually works correctly
-        let remainingKeys = Object.keys(flags).length > 0 ? randomizeSuggestions(flags) : [];
-        return [key, ...remainingKeys];
-      };
-    }
-    return null;
   }
 
   function handleNextQuestion() {
